@@ -14,7 +14,8 @@
     "@types/node": "^16.11.24",
     "framer-motion": "^5.0.0",
     "gh-pages": "^3.2.3",
-    "react": "^17.0.2",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
     "react-helmet": "^6.1.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
@@ -29,17 +30,10 @@
     "eject": "craco eject"
   },
   "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "extends": ["react-app", "react-app/jest"]
   },
   "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
+    "production": [">0.2%", "not dead", "not op_mini all"],
     "development": [
       "last 1 chrome version",
       "last 1 firefox version",
@@ -47,9 +41,8 @@
     ]
   },
   "devDependencies": {
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-router-dom": "^5.3.3",
-    "react-dom": "^17.0.2"
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@types/react-router-dom": "^5.3.3"
   }
 }
